+++ packages/typescript-react-native-resolver/src/host.ts
@@ -11,13 +11,7 @@ import {
   ExtensionsJSON,
   ExtensionsTypeScript,
 } from "./extension";
-import {
-  changeModuleResolutionHostToLogFileSystemReads,
-  logModuleBegin,
-  logModuleEnd,
-  ResolverLog,
-  ResolverLogMode,
-} from "./log";
+import { isTraceEnabled, logModuleBegin, logModuleEnd } from "./log";
 import { createReactNativePackageNameReplacer } from "./react-native-package-name";
 import { resolveFileModule, resolvePackageModule } from "./resolve";
 
@@ -27,18 +21,11 @@ import type { ResolverContext, ModuleResolutionHostLike } from "./types";
  * Change the TypeScript `CompilerHost` or `LanguageServiceHost` so it makes
  * use of react-native module resolution.
  *
- * This includes binding the `trace` method to a react-native trace logger.
- * The logger is active when the compiler option `traceResolution` is true, or
- * when react-native error tracing is enabled. All file and directory reads
- * are logged, making it easy to see what the resolver is doing.
- *
  * @param host Compiler host or language service host
  * @param options Compiler options
  * @param platform Target platform
  * @param platformExtensionNames Optional list of platform file extensions, from highest precedence (index 0) to lowest. Example: `["ios", "mobile", "native"]`.
  * @param disableReactNativePackageSubstitution Flag to prevent substituting the module name `react-native` with the target platform's out-of-tree NPM package implementation. For example, on Windows, devs expect `react-native` to implicitly refer to `react-native-windows`.
- * @param traceReactNativeModuleResolutionErrors Flag to enable trace logging when a resolver error occurs. All messages involved in the failed module resolution are aggregated and logged.
- * @param traceResolutionLog Optional file to use for logging trace message. When not present, log messages go to the console.
  */
 export function changeHostToUseReactNativeResolver({
   host,
@@ -46,16 +33,12 @@ export function changeHostToUseReactNativeResolver({
   platform,
   platformExtensionNames,
   disableReactNativePackageSubstitution,
-  traceReactNativeModuleResolutionErrors,
-  traceResolutionLog,
 }: {
   host: ts.CompilerHost | ts.LanguageServiceHost;
   options: ts.ParsedCommandLine["options"];
   platform: string;
   platformExtensionNames: string[] | undefined;
   disableReactNativePackageSubstitution: boolean;
-  traceReactNativeModuleResolutionErrors: boolean;
-  traceResolutionLog: string | undefined;
 }): void {
   // Ensure that optional methods have an implementation so they can be hooked
   // for logging, and so we can use them in the resolver.
@@ -63,25 +46,10 @@ export function changeHostToUseReactNativeResolver({
   host.realpath = host.realpath ?? ts.sys.realpath;
   host.getDirectories = host.getDirectories ?? ts.sys.getDirectories;
 
-  let mode = ResolverLogMode.Never;
-  if (options.traceResolution) {
-    mode = ResolverLogMode.Always;
-  } else if (traceReactNativeModuleResolutionErrors) {
-    mode = ResolverLogMode.OnFailure;
-  }
-  const log = new ResolverLog(mode, traceResolutionLog);
-  if (mode !== ResolverLogMode.Never) {
-    host.trace = log.log.bind(log);
-    changeModuleResolutionHostToLogFileSystemReads(
-      host as ModuleResolutionHostLike
-    );
-  }
-
   const context: ResolverContext = {
     host: host as ModuleResolutionHostLike,
     options,
     disableReactNativePackageSubstitution,
-    log,
     platform,
     platformExtensions: [platform, ...(platformExtensionNames || [])].map(
       (e) => `.${e}` // prepend a '.' to each name to make it a file extension
@@ -89,8 +57,7 @@ export function changeHostToUseReactNativeResolver({
     replaceReactNativePackageName: createReactNativePackageNameReplacer(
       host.getCurrentDirectory(),
       platform,
-      disableReactNativePackageSubstitution,
-      log
+      disableReactNativePackageSubstitution
     ),
   };
 
@@ -164,21 +131,29 @@ export function resolveModuleNames(
   _reusedNames: string[] | undefined,
   _redirectedReference?: ts.ResolvedProjectReference
 ): (ts.ResolvedModuleFull | undefined)[] {
-  const { options, log, replaceReactNativePackageName } = context;
+  const { options, host, replaceReactNativePackageName } = context;
   const resolutions: (ts.ResolvedModuleFull | undefined)[] = [];
 
+  const traceEnabled = isTraceEnabled(host, options);
+
   for (const moduleName of moduleNames) {
-    logModuleBegin(log, moduleName, containingFile);
+    logModuleBegin(host, options, moduleName, containingFile);
 
     const finalModuleName = replaceReactNativePackageName(moduleName);
+    if (finalModuleName !== moduleName && traceEnabled) {
+      host.trace(
+        `Substituting module '${moduleName}' with '${finalModuleName}'.`
+      );
+    }
 
     //  First, try to resolve the module to a TypeScript file. Then, fall back
     //  to looking for a JavaScript file. Finally, if JSON modules are allowed,
     //  try resolving to one of them.
-    log.log(
-      "Searching for module '%s' with target file type 'TypeScript'",
-      finalModuleName
-    );
+    if (traceEnabled) {
+      host.trace(
+        `Searching for module '${finalModuleName}' with target file type 'TypeScript'`
+      );
+    }
     let module = resolveModuleName(
       context,
       finalModuleName,
@@ -186,10 +161,11 @@ export function resolveModuleNames(
       ExtensionsTypeScript
     );
     if (!module) {
-      log.log(
-        "Searching for module '%s' with target file type 'JavaScript'",
-        finalModuleName
-      );
+      if (traceEnabled) {
+        host.trace(
+          `Searching for module '${finalModuleName}' with target file type 'JavaScript'`
+        );
+      }
       module = resolveModuleName(
         context,
         finalModuleName,
@@ -197,10 +173,11 @@ export function resolveModuleNames(
         ExtensionsJavaScript
       );
       if (!module && options.resolveJsonModule) {
-        log.log(
-          "Searching for module '%s' with target file type 'JSON'",
-          finalModuleName
-        );
+        if (traceEnabled) {
+          host.trace(
+            `Searching for module '${finalModuleName}' with target file type 'JSON'`
+          );
+        }
         module = resolveModuleName(
           context,
           finalModuleName,
@@ -211,7 +188,7 @@ export function resolveModuleNames(
     }
 
     resolutions.push(module);
-    logModuleEnd(log, options, moduleName, module);
+    logModuleEnd(host, options, moduleName, module);
   }
 
   return resolutions;
@@ -247,24 +224,11 @@ export function resolveTypeReferenceDirectives(
   _compilerOptions?: ts.CompilerOptions,
   containingFileMode?: ts.SourceFile["impliedNodeFormat"]
 ): (ts.ResolvedTypeReferenceDirective | undefined)[] {
-  const { host, options, log } = context;
+  const { host, options } = context;
 
   const resolutions: (ts.ResolvedTypeReferenceDirective | undefined)[] = [];
 
   for (const typeDirectiveName of typeDirectiveNames) {
-    // TypeScript only emits trace messages when traceResolution is enabled.
-    // Our code, however, can emit even when traceResolution is disabled. We
-    // have the "emit only errors" mode. This includes the file-system reads
-    // we hooked on the module-resolution host.
-    //
-    // We don't want to see those file-system read messages without the
-    // larger context of TypeScript's trace messages. So, for type directives,
-    // we can only log successes AND failures when traceResolution is enabled.
-    //
-    if (options.traceResolution) {
-      log.begin();
-    }
-
     const name =
       typeof typeDirectiveName === "string"
         ? typeDirectiveName
@@ -281,13 +245,6 @@ export function resolveTypeReferenceDirectives(
       );
 
     resolutions.push(directive);
-    if (options.traceResolution) {
-      if (directive) {
-        log.endSuccess();
-      } else {
-        log.endFailure();
-      }
-    }
   }
 
   return resolutions;