+++ packages/typescript-react-native-resolver/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log - @rnx-kit/typescript-react-native-resolver
 
+## 0.3.0
+
+### Minor Changes
+
+- 94aeb460: BREAKING: update the public method 'changeHostToUseReactNativeResolver'. Remove
+  trace logging parameters. From this version onward, standard TypeScript module
+  resolution tracing applies. Set the compiler option 'traceResolution' to true,
+  and you will see trace messages appear in the console. Logging to a file is no
+  longer supported.
+
 ## 0.2.3
 
 ### Patch Changes